split app.rs into feature submodules
